fix prob bigger than 1 and write double in dbs

+++ src/jzombies/Human.java
@@ -67,7 +67,7 @@ public class Human {
       space.moveByVector(this, 2, angle, 0);
       myPoint = space.getLocation(this);
       grid.moveTo(this, (int) myPoint.getX(), (int) myPoint.getY());
-      Database.updatePoint(name, (int) myPoint.getX(), (int) myPoint.getY());
+      Database.updatePoint(name,myPoint.getX(), myPoint.getY());
       energy--;
     }
   }
